Updated axios and added react-router-dom
Upgraded axios to version 1.6.8 to incorporate the latest fixes and improvements. Introduced react-router-dom@6.22.3 to the project's dependencies for handling routing in React applications. This addition requires adjustments in package-lock.json to include the package and its dependencies, ensuring compatibility and correct version resolution. The upgrade of axios also involves updating follow-redirects dependency for better request handling.

Refs: #40